+++ src/24/outline/ArrowDownRight.js
@@ -1,0 +1,16 @@
+import * as React from 'react'
+const SvgArrowDownRight = (props) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={24}
+    height={24}
+    fill="none"
+    {...props}
+  >
+    <path
+      fill="currentColor"
+      d="M5.22 7.28a.75.75 0 0 1 1.06-1.06L17.5 17.44V9.75a.75.75 0 1 1 1.5 0v9.5a.746.746 0 0 1-.75.75h-9.5a.75.75 0 1 1 0-1.5h7.69L5.22 7.28Z"
+    />
+  </svg>
+)
+export default SvgArrowDownRight