+++ src/24/solid/ArrowLeft.js
@@ -1,0 +1,16 @@
+import * as React from 'react'
+const SvgArrowLeft = (props) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={24}
+    height={24}
+    fill="none"
+    {...props}
+  >
+    <path
+      fill="currentColor"
+      d="M11.771 5.04a.75.75 0 1 0-1.042-1.08l-7.5 7.25a.75.75 0 0 0 0 1.08l7.5 7.25a.751.751 0 0 0 1.042-1.08L5.605 12.5H20.25a.75.75 0 1 0 0-1.5H5.605l6.166-5.96Z"
+    />
+  </svg>
+)
+export default SvgArrowLeft