+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.3] - 2023-11-02
+
+### Added
+
+- Added Arrow Down Left icons.
+- Added Arrow Down Right icons.
+- Added Arrow Down icons.
+- Added Arrow Left icons.
+- Added Arrow Right icons.
+- Added Arrow Up Left icons.
+- Added Arrow Up Right icons.
+- Added Arrow Up icons.
+
 ## [0.1.2] - 2023-11-02
 
 ### Added